Move scan overrides into shim
